src/requirements.txt
setuptools==65.5.1
# metrics
aiohttp==3.8.6
aiodns==2.0.0
mmh3==2.5.1
PyJWT==2.4.0
cryptography==41.0.4
PyYAML==5.4
# logs
google-cloud-pubsub==2.17.1
python-dateutil==2.8.1
jmespath~=0.10.0
pytz==2022.2.1
ciso8601